Update a few scripts, add tool to scale ZTT in pre-made datacards, add the q/g variables to the sync trees

+++ Analysis/HiggsTauTau/scripts/makeDatacards.py
@@ -30,8 +30,18 @@ parser.add_option("-c", "--channels", dest="channels", type='string', action='ca
                   help="A comma separted list of channels to process.  Supported channels: %(CHANNELS)s" % vars())
 parser.add_option("--central", dest="central", action='store_true', default=False,
                   help="Only process the central energy scale inputs.")
+parser.add_option("--mvis", dest="mvis", action='store_true', default=False,
+                  help="Only make inputs for visible mass, no svfit.")
 parser.add_option("-e", dest="energy", type='string', default='8',
                   help="The C.O.M. energy is written into the datacard name, default is 8")
+parser.add_option("--svfit_plot", dest="svfit_plot", type='string', default='m_sv_sm_fine',
+                  help="Override the default svfit plot")
+parser.add_option("--svfit_vbf_plot", dest="svfit_vbf_plot", type='string', default='m_sv_sm',
+                  help="Override the default svfit vbf category plot")
+parser.add_option("--mvis_plot", dest="mvis_plot", type='string', default='m_vis_sm_fine',
+                  help="Override the default mvis plot")
+parser.add_option("--mvis_vbf_plot", dest="mvis_vbf_plot", type='string', default='m_vis_sm',
+                  help="Override the default mvis vbf category plot")
 
 
 (options, args) = parser.parse_args()
@@ -65,35 +75,42 @@ CFG=options.config
 folder = options.input
 COM = options.energy
 
+svfit_plot=options.svfit_plot
+svfit_vbf_plot=options.svfit_vbf_plot
+mvis_plot=options.mvis_plot
+mvis_vbf_plot=options.mvis_vbf_plot
+
 
 for ch in channels:
-  for sc in scales:
-    CATS=[ '8','3','2','1','0' ]
-    for cat in CATS:
-      os.system('./bin/HiggsTauTauPlot3 --cfg=%(CFG)s  --tau_scale_mode=%(sc)s --channel=%(ch)s --rebin=1'
-        ' --method=%(cat)s --plot_name="m_sv_sm_fine"  --x_axis_label="m_{#tau#tau} [GeV]"'
-        ' --blind=false --x_blind_min=100 --x_blind_max=160 --make_datacard=true --norm_bins=true --verbose=false'
-        ' --paramfile=%(PARAMS)s --folder=%(folder)s' % vars())
-    CATS=[ '5' ]
-    for cat in CATS:
-      os.system('./bin/HiggsTauTauPlot3 --cfg=%(CFG)s  --tau_scale_mode=%(sc)s --channel=%(ch)s --rebin=1'
-        ' --method=%(cat)s --plot_name="m_sv_sm"  --x_axis_label="m_{#tau#tau} [GeV]"'
-        ' --blind=false --x_blind_min=100 --x_blind_max=160 --make_datacard=true --norm_bins=true --verbose=false'
-        ' --paramfile=%(PARAMS)s --folder=%(folder)s' % vars())
-  os.system('hadd -f htt_%(ch)s.inputs-sm-%(COM)sTeV%(output)s.root datacard_*.root' % vars())
-  os.system('rm datacard_*.root')
+
+  if not options.mvis:
+    for sc in scales:
+      CATS=[ '8','3','2','1','0' ]
+      for cat in CATS:
+        os.system('./bin/HiggsTauTauPlot3 --cfg=%(CFG)s  --tau_scale_mode=%(sc)s --channel=%(ch)s --rebin=1'
+          ' --method=%(cat)s --plot_name="%(svfit_plot)s"  --x_axis_label="m_{#tau#tau} [GeV]"'
+          ' --blind=false --x_blind_min=100 --x_blind_max=160 --make_datacard=true --norm_bins=true --verbose=false'
+          ' --paramfile=%(PARAMS)s --folder=%(folder)s' % vars())
+      CATS=[ '5' ]
+      for cat in CATS:
+        os.system('./bin/HiggsTauTauPlot3 --cfg=%(CFG)s  --tau_scale_mode=%(sc)s --channel=%(ch)s --rebin=1'
+          ' --method=%(cat)s --plot_name="%(svfit_vbf_plot)s"  --x_axis_label="m_{#tau#tau} [GeV]"'
+          ' --blind=false --x_blind_min=100 --x_blind_max=160 --make_datacard=true --norm_bins=true --verbose=false'
+          ' --paramfile=%(PARAMS)s --folder=%(folder)s' % vars())
+    os.system('hadd -f htt_%(ch)s.inputs-sm-%(COM)sTeV%(output)s.root datacard_*.root' % vars())
+    os.system('rm datacard_*.root')
 
   for sc in scales:
     CATS=[ '8','3','2','1','0' ]
     for cat in CATS:
       os.system('./bin/HiggsTauTauPlot3 --cfg=%(CFG)s  --tau_scale_mode=%(sc)s --channel=%(ch)s --rebin=1'
-        ' --method=%(cat)s --plot_name="m_vis_sm_fine"  --x_axis_label="m_{#tau#tau} [GeV]"'
+        ' --method=%(cat)s --plot_name="%(mvis_plot)s"  --x_axis_label="m_{#tau#tau} [GeV]"'
         ' --blind=false --x_blind_min=100 --x_blind_max=160 --make_datacard=true --norm_bins=true --verbose=false'
         ' --paramfile=%(PARAMS)s --folder=%(folder)s' % vars())
     CATS=[ '5' ]
     for cat in CATS:
       os.system('./bin/HiggsTauTauPlot3 --cfg=%(CFG)s  --tau_scale_mode=%(sc)s --channel=%(ch)s --rebin=1'
-        ' --method=%(cat)s --plot_name="m_vis_sm"  --x_axis_label="m_{#tau#tau} [GeV]"'
+        ' --method=%(cat)s --plot_name="%(mvis_vbf_plot)s"  --x_axis_label="m_{#tau#tau} [GeV]"'
         ' --blind=false --x_blind_min=100 --x_blind_max=160 --make_datacard=true --norm_bins=true --verbose=false'
         ' --paramfile=%(PARAMS)s --folder=%(folder)s' % vars())
   os.system('hadd -f htt_%(ch)s.inputs-sm-%(COM)sTeV-mvis%(output)s.root datacard_*.root' % vars())
